Added a bootstrap configuration for webpack. Converted the bootstrapper.js file to a ts file.
