+++ index.html
@@ -18,7 +18,7 @@
 		<script src="/node_modules/moment/moment.js"></script>
 		<script src="/node_modules/typescript-angular-utilities/output/utilities.js"></script>
 		<script src="/output/components.js"></script>
-		<script src="/bootstrapper/bootstrapper.js"></script>
+		<script src="/bootstrapper/app.js"></script>
 		<script src="/bootstrapper/messageLogBootstrapper.js"></script>
 		<script src="/bootstrapper/cardContainerBootstrapper.js"></script>
 		<script src="/bootstrapper/text.js"></script>
